src/test/java/com/oopexercises1/oopexercises/collections/MorseCodeTest.java
package com.oopexercises1.oopexercises.collections;

import com.oopexercises.collections.MorseCode;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertThrows;

class MorseCodeTest {

    @Test
    void morseCode() {
        assertEquals(".... . ._.. ._.. ___", MorseCode.morseCode("hello"));
        assertEquals(".__ ___ ._. ._.. _..", MorseCode.morseCode("world"));
        assertEquals(".__ _____ ._. ._.. _..", MorseCode.morseCode("w0rld"));
        assertEquals("", MorseCode.morseCode(""));
        assertThrows(IllegalArgumentException.class, () -> MorseCode.morseCode("w$rld"));
    }
}